Move `pad_inputs` into padding module
